pyproject.toml
[tool.poetry]
name = "opencicd"
version = "0.0.1"
description = "OpenSource CI/CD tool"
authors = ["Ruslan Khyurri <ruslan_khyurri@epam.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.13.0"
black = "^21.6b0"
mypy = "^0.902"
pylint = "^2.8.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
